Add structlog-based logger with context binding
This adds a structlog-based logger with a middleware, bind_logger, that
binds request context to the logger. sqrbot.logging stores this logger
as a contextlib.ContextVar so that it is local to each response task.

The logger is also available from the request object in the handler.
(logger = request['logger'])

The logger is configured in sqrbot.app.configure_logging(), which
configures the logger with JSON formatting in production and coloured
key-value formatting in development.

The logger is named after the api.lsst.codes/loggerName configuration
field, if possible.