Add Kubernetes deployment
This is an MVP deployment for Kubernetes that contains all the basic
elements: deployment, configmaps, service, and ingress.

The ingress is interesting because it's meant to work with the
nginx-ingress. Essentially, this Ingress resource triggers a
configuration of the nginx controller.

Note that this MVP does not include an nginx reverse proxy container
just yet. That's the next step.