Use signing secret to verify Slack request
This adds a _verify_request function that implements the signing secret
algorithm to verify the authenticity of a Slack request. This is more
secure than the token method.

The implementation is based on the docs at
https://api.slack.com/docs/verifying-requests-from-slack as well as the
Flask-based library
https://github.com/slackapi/python-slack-events-api/blob/master/slackeventsapi/server.py
This implementation is simplified since it's Python 3-only.

Since _verify_request is called at the top of the handler, other code
such as the challenge handler not longer need to worry about verify
tokens.