Add system for generating event schemas
In the Slack Events API, all events share the same wrapper
(https://api.slack.com/types/event), but the events themselves have
different schemas for the "events" field. The load_event_schema function
merges a schema specific to each event into the Event type wrapper. The
name of the event schema is also updated with the specific name of the
event. The message event is included as an initial proof-of-concept.

There are a couple additions to the schema:

Additional notes:

- It turns out that the main schema is for message itself, and that
  things like message.im or message.channel are just different events you
  can subscribe to.

- Add channel_type field to message schema

  Messages usually have a channel_type field that can take the value of
  "channel", "im", "mpim", or "group"

- Added both authed_teams and authed_users to the wrapper's schema. The
  Slack JSON schema suggested that authed_users is required, but their
  examples don't include it; they only include authed_teams. I'll include
  both fields, but make them optional (defaulting to null).