Add stagingVersion config for schema name suffix
The sqrbot-jr/stagingVersion configuration can be set to indicate that
the app is being run in a staging environment. In particular, this means
that schemas and topics aren't production. To isolate these staging
events from productions apps, the stagingVersion gets added as a suffix
to the names of schemas (which become the names of subjects) and will
also get applied to the names of topics.