Create Kafka topics on start up
By check and if necessary creating topics on startup, we can ensure that
the topics get created with the right partitioning and replication
factor (we don't want to leave this to chance).

This adds the confluent-kafka package as a direct dependency. Perhaps in
the future we'll add a Kafka Broker API wrapper to Kafkit.