Use underscores to append versions to names
It turns out that hypens (-) are not permitted in Avro names, see:

http://avro.apache.org/docs/current/spec.html#names

So instead of appending the major version to the schema name like
"messages-v1", we now write the name as "messages_v1".

The same also applies to staging version names. For example:

"messages-v1-dev" is now "messages_v1_dev".