Add key schema for interaction events
Interactions are processed by the handler for the /interaction endpoint.
These can be message actions, dialogs/modals, interactive messages.

The goal with this key is to partition these events according to who is
doing the action. This way the events made by a given user are processed
in sequence, while there is opportunity to process events made by
different users in parallel.