Convert app to FastAPI
This converts the app's infrastructure and handlers into FastAPI. In the
conversion we've temporarily dropped publishing Kafka messages, through
we've roughed in a Slack service with handlers for messages and
interactions. The Slack payloads are defined in models in the
squarebot.domain.slack module. Some detail is lost on the interaction
payloads and this is partly because the Slack API seems to have
modernized w.r.t. interaction payloads so we'll need to study this more.