Add nox run session
This session runs the app under uvicorn. Note it doesn't work at the
moment because a Kafka cluster is expected on app startup. This session
is more a proof of concept to replace the makefile/tox run environment.
A fully working replacement will need a way to set up external services
as docker containers, likely.