Set up FastStream-based Kafka producer
This commit re-implements Squarebot's Kafka producer as a
FastStream-based producer that directly uses Pydantic models serialized
to JSON rather than the original system of including the Confluent
Schema Registry in message encoding. Using Pydantic models directly
makes sense for Squarebot because the models are being distributed
through the rubin-squarebot PyPI package. Further using FastStream
provides some niceities around running Kafka in the context of a FastAPI
app.

For each topic, we maintain a Publisher instance that runs for the
lifespan of the app. We keep these pubilshers in a ProcessContext
object, which is available through the RequestContext dependency, and
included in the new Factory. The Factory provides instances of the Slack
service to requests handlers.

For testing, we now rely on end-to-end testing with a local Kafka
broker. To deploy this broker, run docker compose with the
kafka-compose.yaml definitions.