Add bot_message support for Slack messages
Add support for Slack messages with `subtype` of `bot_message`. This is
implemented in the Slack message model `SlackMessageEventContent` as a
new `subtype` field that now takes a `SlackMessageSubtype` enum. The
Kafka representation of this message, `SquarebotSlackMessageValue` has
been updated to now set the `user` attribute to either the user's ID or
the bot's ID as appropriate. A new flag, `is_bot` indicates if the
`user` field reflects a user or bot (app integration) identifier.