Support getting Slack text content in attachments
The `SquarebotSlackMessageValue` Kafka message model's `text` field
now includes the combined content of the message and its attachments.
This makes it easy for consumers to process message content without
needing to check for attachments and Block Kit fields. For consumers
that _are_ sensitive to the exact structure of the message, the
`slack_event` field is still available, which is the original JSON
representation of the Slack message, which can be parsed into a
`SlackMessageEvent` object.